Thu, Sep 7, 2023, 10:00 AM +02:00

+++ ProblemSet4/filter-more/helpers.c
@@ -8,7 +8,8 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int avg = ceil((image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed) / 3);
+            int avg = ceil((float)(image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed) / 3.0F);
+
             if (avg > 0)
             {
                 image[i][j].rgbtBlue = avg;
